EES-3828 Final fixes before making the PR
