+++ src/GovUk.Education.ExploreEducationStatistics.Admin.Tests/Services/ContentServiceTests.cs
@@ -644,117 +644,6 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Tests.Services
             }
         }
 
-        [Fact]
-        public async Task GetAvailableDataBlocks()
-        {
-            var release = new Release();
-
-            var availableDataBlock1Id = Guid.NewGuid();
-            var availableDataBlock2Id = Guid.NewGuid();
-
-            var unavailableDataBlock1Id = Guid.NewGuid();
-            var keyStat = new KeyStatisticDataBlock
-            {
-                Release = release,
-                DataBlockId = unavailableDataBlock1Id,
-            };
-
-            var releaseContentBlocks = new List<ReleaseContentBlock>
-            {
-                new()
-                {
-                    Release = release,
-                    ContentBlock = new DataBlock
-                    {
-                        Id = unavailableDataBlock1Id, // unavailable because has key stat
-                        Name = "Unavailable 1",
-                        ContentSectionId = null,
-                    },
-                },
-                new()
-                {
-                    Release = release,
-                    ContentBlock = new DataBlock
-                    {
-                        Id = availableDataBlock1Id,
-                        Name = "Available 1",
-                        ContentSection = null,
-                    },
-                },
-                new()
-                {
-                    Release = release,
-                    ContentBlock = new HtmlBlock(),
-                },
-                new()
-                {
-                    Release = release,
-                    ContentBlock = new DataBlock
-                    {
-                        Id = availableDataBlock2Id,
-                        Name = "Available 2",
-                        ContentSection = null,
-                    },
-                },
-                new()
-                {
-                    Release = release,
-                    ContentBlock = new DataBlock
-                    {
-                        Name = "Unavailable 2", // because has content section
-                        ContentSection = new ContentSection(),
-                    },
-                },
-                new()
-                {
-                    Release = new Release(),
-                    ContentBlock = new DataBlock
-                    {
-                        Name = "Unavailable 3", // because different release
-                        ContentSection = null,
-                    }
-                },
-            };
-
-            var contentDbContextId = Guid.NewGuid().ToString();
-            await using (var contentDbContext = InMemoryApplicationDbContext(contentDbContextId))
-            {
-                await contentDbContext.ReleaseContentBlocks.AddRangeAsync(releaseContentBlocks);
-                await contentDbContext.KeyStatisticsDataBlock.AddRangeAsync(keyStat);
-                await contentDbContext.SaveChangesAsync();
-            }
-
-            await using (var contentDbContext = InMemoryApplicationDbContext(contentDbContextId))
-            {
-                var service = SetupContentService(contentDbContext: contentDbContext);
-                var result = await service.GetAvailableDataBlocks(
-                    release.Id);
-
-                var availableDataBlocks = result.AssertRight();
-
-                Assert.Equal(2, availableDataBlocks.Count);
-
-                Assert.Equal(availableDataBlock1Id, availableDataBlocks[0].Id);
-                Assert.Equal("Available 1", availableDataBlocks[0].Name);
-                Assert.Equal(availableDataBlock2Id, availableDataBlocks[1].Id);
-                Assert.Equal("Available 2", availableDataBlocks[1].Name);
-            }
-        }
-
-        [Fact]
-        public async Task GetAvailableDataBlocks_NoRelease()
-        {
-            var contentDbContextId = Guid.NewGuid().ToString();
-            await using (var contentDbContext = InMemoryApplicationDbContext(contentDbContextId))
-            {
-                var service = SetupContentService(contentDbContext: contentDbContext);
-                var result = await service.GetAvailableDataBlocks(
-                    Guid.NewGuid());
-
-                result.AssertNotFound();
-            }
-        }
-
         private static ContentService SetupContentService(
             ContentDbContext contentDbContext,
             IPersistenceHelper<ContentDbContext> persistenceHelper = null,
