+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Controllers/Api/DataBlocksController.cs
@@ -5,6 +5,7 @@ using GovUk.Education.ExploreEducationStatistics.Admin.Services;
 using GovUk.Education.ExploreEducationStatistics.Admin.Services.Interfaces;
 using GovUk.Education.ExploreEducationStatistics.Admin.ViewModels;
 using GovUk.Education.ExploreEducationStatistics.Common.Extensions;
+using GovUk.Education.ExploreEducationStatistics.Content.Model;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
@@ -72,5 +73,14 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Controllers.Api
                 .Update(id, dataBlock)
                 .HandleFailuresOrOk();
         }
+
+        [HttpGet("release/{releaseId}/data-blocks/available")]
+        public async Task<ActionResult<List<DataBlock>>> GetAvailableDataBlocks(Guid releaseId)
+        {
+            return await _dataBlockService
+                .GetAvailableDataBlocks(releaseId)
+                .HandleFailuresOrOk();
+        }
+
     }
 }
