+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Services/DataBlockService.cs
@@ -376,6 +376,31 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Services
                 .CreateCacheKeyForDataBlock(releaseId, dataBlockId)
                 .OnSuccessVoid(_cacheService.DeleteItem);
         }
+
+        public async Task<Either<ActionResult, List<DataBlock>>> GetAvailableDataBlocks(Guid releaseId)
+        {
+            return await _persistenceHelper.CheckEntityExists<Release>(releaseId)
+                .OnSuccess(_userService.CheckCanViewRelease)
+                .OnSuccess(async release =>
+                {
+                    var keyStatDataBlockIds = _context.KeyStatisticsDataBlock
+                        .Where(ks => ks.ReleaseId == release.Id)
+                        .Select(ks => ks.DataBlockId)
+                        .ToList();
+
+                    return await _context
+                        .ReleaseContentBlocks
+                        .Include(releaseContentBlock => releaseContentBlock.ContentBlock)
+                        .Where(releaseContentBlock => releaseContentBlock.ReleaseId == release.Id)
+                        .Select(releaseContentBlock => releaseContentBlock.ContentBlock)
+                        .OfType<DataBlock>()
+                        .Where(dataBlock =>
+                            dataBlock.ContentSectionId == null
+                            && !keyStatDataBlockIds.Contains(dataBlock.Id))
+                        .OrderBy(contentBlock => contentBlock.Name)
+                        .ToListAsync();
+                });
+        }
     }
 
     public class DependentDataBlock
