+++ src/explore-education-statistics-admin/src/components/editable/EditableKeyStatText.tsx
@@ -20,9 +20,9 @@ export interface EditableKeyStatTextProps {
 }
 
 const EditableKeyStatText = ({
+  keyStat,
   isEditing = false,
   isReordering = false,
-  keyStat,
   testId = 'keyStat',
   onRemove,
   onSubmit,
