+++ src/explore-education-statistics-admin/src/services/dataBlockService.ts
@@ -77,6 +77,10 @@ const dataBlockService = {
       {},
     );
   },
+
+  getAvailableDataBlocks(releaseId: string): Promise<DataBlock[]> {
+    return client.get(`/release/${releaseId}/data-blocks/available`);
+  },
 };
 
 export default dataBlockService;
