+++ tests/robot-tests/tests/libs/common.robot
@@ -750,7 +750,7 @@ user checks publication bullet does not contain link
     ...    xpath://details[@open]//*[text()="${publication}"]/..//a[text()="${link}"]
 
 user checks key stat contents
-    [Arguments]    ${tile}    ${title}    ${value}    ${summary}    ${wait}=${timeout}
+    [Arguments]    ${tile}    ${title}    ${value}    ${trend}    ${wait}=${timeout}
     user waits until element is visible
     ...    css:[data-testid="keyStat"]:nth-of-type(${tile}) [data-testid="keyStat-title"]    ${wait}
     user waits until element contains    css:[data-testid="keyStat"]:nth-of-type(${tile}) [data-testid="keyStat-title"]
@@ -758,15 +758,15 @@ user checks key stat contents
     user waits until element contains    css:[data-testid="keyStat"]:nth-of-type(${tile}) [data-testid="keyStat-value"]
     ...    ${value}
     user waits until element contains
-    ...    css:[data-testid="keyStat"]:nth-of-type(${tile}) [data-testid="keyStat-summary"]    ${summary}
+    ...    css:[data-testid="keyStat"]:nth-of-type(${tile}) [data-testid="keyStat-trend"]    ${trend}
 
 user checks key stat definition
-    [Arguments]    ${tile}    ${definition_summary}    ${definition}
-    user opens details dropdown    ${definition_summary}    css:[data-testid="keyStat"]:nth-of-type(${tile})
+    [Arguments]    ${tile}    ${guidance_title}    ${guidance_text}
+    user opens details dropdown    ${guidance_title}    css:[data-testid="keyStat"]:nth-of-type(${tile})
     user waits until element is visible
-    ...    css:[data-testid="keyStat"]:nth-of-type(${tile}) [data-testid="keyStat-definition"]
+    ...    css:[data-testid="keyStat"]:nth-of-type(${tile}) [data-testid="keyStat-guidanceText"]
     user checks element should contain
-    ...    css:[data-testid="keyStat"]:nth-of-type(${tile}) [data-testid="keyStat-definition"]    ${definition}
+    ...    css:[data-testid="keyStat"]:nth-of-type(${tile}) [data-testid="keyStat-guidanceText"]    ${guidance_text}
 
 user checks page contains radio
     [Arguments]    ${label}
